Allow files to contain rules over multiple lines
Previously, program would panic when trying to convert rules over
multiple lines as the line would not contain '::=' as it would split it
and then try to access the second index.

Now, it will find all lines that contain '::=' and then join any lines
that do not have them onto the last one that did.